Implement --from-github option for repo command (#66)
